treewide: Finalising work
flash_rescue_userspace with FlashRescueBoardApp successfully reflash the
SPI chip. However, userspace reports failures to verify, though this is
incorrect. The flash chip holds the transferred binary (even if it's
broken. Oops, wasn't ready in time).

FlashRescueBoardPei is successfully reloaded into DRAM on Simics (NEM is
impossible, minimal is available). Work now proceeds on porting the
shell app to a PEIM.

So, an interesting finding: the Bus Pirate can support 1 MHz baudrate
between FTDI and PIC. Primitive testing, printing ~50 chars x 256 times
at GMBUS 100K with reduced stall to... 1ms or 2ms, as I recall?
Indicates that time will be greatly improved. The test changes from
4.4s to 1.9s. So, functionality to switch baudrates was implemented
here. Success cannot be determined yet, as 400K was attempted here and
failed.

+++ flash_rescue_userspace/util.h
@@ -4,12 +4,14 @@
 #define UTIL_H
 
 #include <stdbool.h>
+#include <stdint.h>
+#include <termios.h>
 
 #define TO_PERCENTAGE(val, total) (100 - (((total - val) * 100) / total))
 
 int serial_open(char *dev, speed_t baud);
-void serial_fifo_write(void *data, int number_of_bytes);
-void serial_fifo_read(void *data, int number_of_bytes);
+void serial_fifo_write(void *data, size_t number_of_bytes);
+void serial_fifo_read(void *data, size_t number_of_bytes);
 void bp_switch_baudrate_generator(bool to_high_speed);
 void bp_exit(void);
 void sig_handler(int sig_num);
